+++ src/components/admin/SideMenuAdminDesktop.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { Home, LucideSalad, Bed, UserRound, LogOut, ChevronDown, ChevronUp, SquareArrowLeft, SquareArrowRight } from 'lucide-react';
+import { Home, LucideSalad, Bed, UserRound, LogOut, ChevronDown, ChevronUp, SquareArrowLeft, SquareArrowRight, Mail } from 'lucide-react';
 import styles from './styles/SideMenuAdminDesktop.module.css';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
@@ -21,6 +21,7 @@ export function SideMenuAdminDesktop({ set }: SideMenuAdminDesktopProps) {
     const menuItems = [
         { id: 'home', label: 'Início', icon: Home, href: '/admin/home' },
         { id: 'usuarios', label: 'Usuários', icon: UserRound, href: '/admin/usuarios' },
+        { id: 'comunicados', label: 'Comunicados', icon: Mail, href: '/admin/comunicados' },
         { id: 'refeicoes', label: 'Refeições', icon: LucideSalad,
             submenu: [
                 { id: "lista_de_refeicoes", label: "Lista de refeições", href: "/admin/refeicoes"},